Move `LLMServiceInternal` to a separate file
