+++ src/llm/llmServices/lmStudio/lmStudioService.ts
@@ -5,9 +5,9 @@ import { ChatHistory } from "../chat";
 import {
     GeneratedProofImpl,
     LLMServiceImpl,
-    LLMServiceInternal,
     ProofVersion,
 } from "../llmService";
+import { LLMServiceInternal } from "../llmServiceInternal";
 import { LMStudioModelParams } from "../modelParams";
 
 import { LMStudioModelParamsResolver } from "./lmStudioModelParamsResolver";
