+++ src/test/llm/llmSpecificTestUtils/dummyLLMService.ts
@@ -6,9 +6,9 @@ import {
     ErrorsHandlingMode,
     GeneratedProofImpl,
     LLMServiceImpl,
-    LLMServiceInternal,
     ProofVersion,
 } from "../../../llm/llmServices/llmService";
+import { LLMServiceInternal } from "../../../llm/llmServices/llmServiceInternal";
 import {
     ModelParams,
     modelParamsSchema,
